Move sync logic to be run always
Lift sync logic out of energy_aware_wake_cpu and run that logic before
select_idle_sibling() or energy_aware_wake_cpu().

At high CPU utilization, the scheduler falls back to
select_idle_sibling().  On Marlin, select_idle_sibling() wasn't
respecting the sync flag to force the task to be scheduled on the same
CPU.  This caused the sync flag to be ignored and thrash on binder with
sync enabled.  This fixes the thrash we saw.

Test:  manually ran hwbinder_benchmark
Change-Id: I47d21bd3ef936d8e0980b618145de29d3ef6d480
Signed-off-by: Eliminater74 <eliminater74@gmail.com>