sched: Add back capacity margin in sched_freq_tick_walt
Previously we added capacity margin in tick, before the margin was even
considered in cpufreq. Then we stopped doing this in 807a50d14f because
it was redundant. However now it looks like CFS tasks can suffer as seen
in higher app install times.

This patch adds back the code for the margin addition that was
previously removed for CFS. We will tackle that in a separate bug since
it was not a part of the initial patch that was to fix frequency scaling
for RT. With this the app install times are back to what they used to
be.

Bug: 34077789

Change-Id: Iba8970fd24376f05be1a2e2df53f8ef85d23b463
Signed-off-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>